Move admin integration tests to separate package and tweak Travis config
